Added RAM usage calculations in % value. For now only usage nothing else is important in this project i guess

+++ src/gpu_info/gpu_params/gpu_memory_info.cpp
@@ -7,9 +7,9 @@
 
 
 
-std::vector<unsigned long long> get_gpu_VRAM_info() {
+std::vector<unsigned int> get_gpu_VRAM_info() {
 
-  std::vector<unsigned long long> memory_info;
+  std::vector<unsigned int> memory_info;
 
   nvmlDevice_t device = get_device();
   nvmlMemory_t memory;
@@ -21,9 +21,9 @@ std::vector<unsigned long long> get_gpu_VRAM_info() {
     return {};
   }
   float VRAM_memory_percentage_usage             = 100 * ((float)memory.used / (float)memory.total);
-  unsigned long long total_VRAM_installed_MiB    =     (memory.total / 1024 / 1024);
-  unsigned long long total_VRAM_used_MiB         =     (memory.used / 1024 / 1024);
-  unsigned long long total_VRAM_free_MiB         =     (memory.free / 1204 / 1024);
+  unsigned int total_VRAM_installed_MiB    =     (memory.total / 1024 / 1024);
+  unsigned int total_VRAM_used_MiB         =     (memory.used / 1024 / 1024);
+  unsigned int total_VRAM_free_MiB         =     (memory.free / 1204 / 1024);
   memory_info.emplace_back(total_VRAM_installed_MiB);
   memory_info.emplace_back(total_VRAM_used_MiB);
   memory_info.emplace_back(total_VRAM_free_MiB);
